isctespot-main/server/requirements.txt
Flask==2.3.1
Flask-Cors==3.0.9
mariadb==1.1.5
requests==2.19.1
pycryptodome==3.17
pandas==2.2.2
openpyxl==3.1.5
PyJWT==1.5.0
cryptography==3.3.2

# ----------------------------------------------------
# SEGURANÇA E CRIPTOGRAFIA (MITIGAÇÕES)
# ----------------------------------------------------

# Proposta 1: Hashing de Senhas
bcrypt==4.0.1

# Proposta 2: Criptografia Simétrica (AES-GCM)
pycryptodome==3.17

# Gestão de Segredos (Mitigação de Hardcoded Keys)
python-dotenv==1.0.0